nikkei_autotrade/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
websocket-client==1.8.0
